review: conditional loading of the validate and reset-password views
